Revert "Add bookings"

+++ src/main/java/ru/practicum/shareit/item/exceptions/ItemNotFoundException.java
@@ -4,7 +4,7 @@ public class ItemNotFoundException extends RuntimeException {
 
     private static final String MES_PATTERN = "Item with id %s does not exist.";
 
-    public ItemNotFoundException(Long id) {
+    public ItemNotFoundException(Integer id) {
         super(String.format(MES_PATTERN, id));
     }
 }
